requirements.txt
streamlit==1.35.0
pandas==2.2.2
seaborn==0.13.2
matplotlib==3.8.4
python-dotenv==1.0.1
mysql-connector-python==8.3.0
